Tests/LibWeb/Text/expected/DOM/beforeunload.txt

Before unload event fired
Default prevented: true
